requirements.txt
torch
torchvision
pillow
kornia
transformers
huggingface_hub